refactor: extract `Delete`
