+++ components/Admin/Illustration/Illustration.js
@@ -1,55 +1,16 @@
 import { useState } from 'react'
-import { Image, Button, Table, Modal, message } from 'antd'
-import { useQueryClient } from 'react-query'
+import { Image, Button, Table } from 'antd'
 import { DndProvider } from 'react-dnd'
 import { HTML5Backend } from 'react-dnd-html5-backend'
 import DraggableBodyRow from 'components/Common/DraggableBodyRow'
-import NewTabLink from 'components/Common/NewTabLink'
-import useIllustration from 'components/api/useIllustration'
-import useDeleteIllustration from 'components/api/useDeleteIllustration'
-import { Container, Row, DeleteIcon } from '../Admin.style'
+import { Container, Row } from '../Admin.style'
 import UploadModal from './UploadModal'
+import Delete from './Delete'
 import useRows from './useRows'
 
 function Illustration() {
   const [rows, moveRow, canDrag] = useRows()
-
-  const { queryKey } = useIllustration()
   const [modalVisible, setModalVisible] = useState(false)
-
-  const queryClient = useQueryClient()
-  const { submit } = useDeleteIllustration()
-
-  function deleteIllustration(id, name) {
-    submit(
-      { id, name },
-      {
-        onSuccess: () => {
-          queryClient.setQueryData(queryKey, oldData =>
-            oldData.filter(data => data.id !== id)
-          )
-
-          message.success('刪除成功')
-        },
-        onError: () => {
-          message.error('刪除失敗')
-        },
-      }
-    )
-  }
-
-  function confirm({ id, name, src }) {
-    return () => {
-      Modal.confirm({
-        centered: true,
-        title: '你確認要刪除這個圖片嗎？',
-        content: <NewTabLink href={src}>{name}</NewTabLink>,
-        cancelText: '取消',
-        okText: '刪除',
-        onOk: () => deleteIllustration(id, name),
-      })
-    }
-  }
 
   return (
     <>
@@ -98,9 +59,7 @@ function Illustration() {
               {
                 title: '操作',
                 key: 'action',
-                render: ({ id, name, src }) => (
-                  <DeleteIcon onClick={confirm({ id, name, src })} />
-                ),
+                render: illustration => <Delete {...illustration} />,
               },
             ]}
           />
